+++ backend/testSetup.ts
@@ -1,0 +1,1 @@
+jest.mock('./mempool-config.json', () => ({}), { virtual: true });
